+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
-## v2.17.0 - Novemeber 5th, 2025
+## v2.17.1 - November 11th, 2025
+### Added
+- Option to remove Profile Effects (Animated Overlays) [Profile Customizations]
+
+### Fixed
+- Remove Nameplate not removing nameplates
+- Core toggles triggering a style refresh
+
+## v2.17.0 - November 5th, 2025
 ### Added
 - Support for multiple update channels to prevent breakage for people with certain experiments [Core]
     - Sub-channels will be updated less frequently, so please check the main branch often for the latest updates